Add files via upload
These are data analytics projects where I used SQL to inspect and clean data. The data used is hypothetical data for a country called Maji Ndogo. The intention of the project is to help solve our water crisis here in Maji Ndogo. In the project, I used basic to advanced SQL skills.